Add gun and bullet items with automated constructors and entity. Bullet items can have effects.
This commit introduces a new gun and bullet system to the game. The gun and bullet items have been added with simple, automated constructors to streamline their creation. Additionally, a new bullet entity has been added to handle bullets once they are fired from the gun.

The bullet items can also now have effects applied to them, which can be set when the bullet is created. This allows for greater customization and strategy when using the gun.